requirements.txt
imageio>=2.9.0
opencv-python>=4.2.0.34
Pillow>=8.3.1
PySimpleGUI>=4.45.0
numpy>=1.21.0
pytesseract>=0.3.8
keras>=2.6.0